refactor: change limit param to be usize and truncate wallpaper vecs

+++ wally-providers/src/providers.rs
@@ -10,7 +10,7 @@ pub mod wallhaven;
 #[async_trait]
 pub trait WallpaperProvider {
     /// Query the wallpaper provider for a list of wallpaper source urls
-    async fn list(&self, limit: u32) -> anyhow::Result<Vec<Url>>;
+    async fn list(&self, limit: usize) -> anyhow::Result<Vec<Url>>;
 
     /// Retreive a random wallpaper url
     async fn random(&self) -> anyhow::Result<Url>;
